+ on_zagadka & colly scrapper

+++ app/output/view_message_data.go
@@ -2,7 +2,7 @@ package output
 
 // ViewMessageData - simple message data to be shown by IView
 type ViewMessageData struct {
-	ReplyToMsgID int64
+	ReplyToMsgID int
 	ChatID       int64
 	Text         string
 }
